Address PR feedback: use real generateIframeHtml() in Playwright security tests
- Extract generateIframeHtml, JSXGRAPH_VERSION, and JSXGRAPH_CDN_URL into
  GraphSandbox.iframe.ts so Playwright specs can import without JSX issues
- Rewrite graph-sandbox-security.spec.ts to build test iframes from the real
  generateIframeHtml() output instead of a hardcoded CSP copy
- Use page.route() for deterministic network-blocking assertions
- Keep parent-access test tied to the real sandbox="allow-scripts" + CSP
- Run npm run test:ui (2 passed), npm run build, npm run test:smoke:compose
  (3 passed), and full unit suite (304 passed)
