+++ frontend/tests/graph-sandbox-security.spec.ts
@@ -1,28 +1,23 @@
-import { expect, test, type Page } from "@playwright/test";
+import { expect, test } from "@playwright/test";
 import { APP_BASE } from "./helpers";
+import { generateIframeHtml } from "../src/components/GraphSandbox.iframe";
 
 test.use({ baseURL: APP_BASE });
 
 test.describe("GraphSandbox Security", () => {
-  let testPage: Page;
-
-  test.beforeEach(async ({ page }) => {
-    testPage = page;
-  });
-
   test("should block outbound fetch requests from within the iframe", async ({ page }) => {
-    // Track all network requests
-    const requestedUrls: string[] = [];
-    page.on("request", (request) => {
-      requestedUrls.push(request.url());
+    // Use page.route for deterministic assertion that no request reaches example.com
+    let routeHit = false;
+    await page.route("**/example.com/**", (route) => {
+      routeHit = true;
+      route.abort();
     });
 
-    // Navigate to the app
     await page.goto("/");
 
-    // Inject a test iframe with the same CSP and sandbox as GraphSandbox
-    await page.evaluate(() => {
-      return new Promise((resolve) => {
+    // Inject iframe using real generateIframeHtml() and the postMessage protocol
+    await page.evaluate((html) => {
+      return new Promise<void>((resolve) => {
         const iframe = document.createElement("iframe");
         iframe.id = "test-graph-sandbox-iframe";
         iframe.sandbox = "allow-scripts";
@@ -30,80 +25,49 @@ test.describe("GraphSandbox Security", () => {
         iframe.style.left = "-9999px";
         iframe.style.top = "-9999px";
 
-        // Same CSP as GraphSandbox
-        const iframeHtml = `
-          <!DOCTYPE html>
-          <html>
-          <head>
-            <meta charset="UTF-8">
-            <meta http-equiv="Content-Security-Policy" content="
-              default-src 'none';
-              script-src 'unsafe-eval' 'unsafe-inline' https://cdn.jsdelivr.net/npm/jsxgraph@1.10.0/distrib/jsxgraphcore.js;
-              style-src 'unsafe-inline';
-              img-src data:;
-              connect-src 'none';
-              font-src 'none';
-              frame-src 'none';
-              object-src 'none';
-              media-src 'none';
-              base-uri 'none';
-              form-action 'none';
-            ">
-            <script src="https://cdn.jsdelivr.net/npm/jsxgraph@1.10.0/distrib/jsxgraphcore.js"></script>
-            <script>
-              // Try to make a fetch request
-              fetch('https://example.com/malicious-test-request')
-                .catch(() => {});
+        const blob = new Blob([html], { type: "text/html" });
+        iframe.src = URL.createObjectURL(blob);
 
-              // Also try XMLHttpRequest
-              try {
-                const xhr = new XMLHttpRequest();
-                xhr.open('GET', 'https://example.com/malicious-xhr-request', true);
-                xhr.send();
-              } catch (e) {}
-
-              // Notify parent that we've attempted the requests
-              parent.postMessage({ type: 'test-done' }, '*');
-            </script>
-          </head>
-          <body>
-            <div id="jxgbox"></div>
-          </body>
-          </html>
-        `;
-
-        const blob = new Blob([iframeHtml], { type: "text/html" });
-        const blobUrl = URL.createObjectURL(blob);
-        iframe.src = blobUrl;
-
-        // Wait for the iframe to send the test-done message
         window.addEventListener("message", function handler(event) {
-          if (event.data?.type === "test-done") {
+          if (event.data?.type === "ready") {
             window.removeEventListener("message", handler);
-            resolve(undefined);
+            // Send malicious DSL via postMessage (same protocol as GraphSandbox)
+            iframe.contentWindow?.postMessage(
+              {
+                type: "render",
+                payload: `
+                  fetch('https://example.com/malicious-test-request').catch(() => {});
+                  try {
+                    const xhr = new XMLHttpRequest();
+                    xhr.open('GET', 'https://example.com/malicious-xhr-request', true);
+                    xhr.send();
+                  } catch (e) {}
+                `,
+              },
+              "*"
+            );
+
+            // Wait for any requests to be attempted, then signal completion
+            setTimeout(() => resolve(undefined), 1500);
           }
         });
 
         document.body.appendChild(iframe);
       });
-    });
+    }, generateIframeHtml());
 
-    // Wait a bit for any requests to go through
-    await page.waitForTimeout(1000);
+    // Wait for the iframe test to complete
+    await page.waitForTimeout(2000);
 
     // Verify no requests to example.com were made
-    const hasExampleComRequest = requestedUrls.some(url =>
-      url.includes("example.com")
-    );
-    expect(hasExampleComRequest).toBe(false);
+    expect(routeHit).toBe(false);
   });
 
   test("should block access to parent document from within the iframe", async ({ page }) => {
-    // Navigate to the app
     await page.goto("/");
 
-    // Inject a test iframe to check parent access
-    const result = await page.evaluate(() => {
+    // Inject iframe using real generateIframeHtml() with an appended parent-access probe
+    const result = await page.evaluate((html) => {
       return new Promise((resolve) => {
         const iframe = document.createElement("iframe");
         iframe.id = "test-parent-access-iframe";
@@ -112,32 +76,24 @@ test.describe("GraphSandbox Security", () => {
         iframe.style.left = "-9999px";
         iframe.style.top = "-9999px";
 
-        const iframeHtml = `
-          <!DOCTYPE html>
-          <html>
-          <head>
-            <meta charset="UTF-8">
-            <script>
-              let parentAccessResult: { blocked: boolean; message?: string } = { blocked: false };
+        // Append a script that probes parent access before </body>
+        const probeScript = `
+          <script>
+            (function() {
+              let parentAccessResult = { blocked: false };
               try {
-                // Try to access parent.document
                 const _ = window.parent.document;
-                parentAccessResult = { blocked: false };
               } catch (e) {
                 parentAccessResult = { blocked: true, message: e instanceof Error ? e.message : String(e) };
               }
-
-              // Notify parent of the result
               parent.postMessage({ type: 'parent-access-result', result: parentAccessResult }, '*');
-            </script>
-          </head>
-          <body></body>
-          </html>
+            })();
+          </script>
         `;
+        const modifiedHtml = html.replace("</body>", `${probeScript}</body>`);
 
-        const blob = new Blob([iframeHtml], { type: "text/html" });
-        const blobUrl = URL.createObjectURL(blob);
-        iframe.src = blobUrl;
+        const blob = new Blob([modifiedHtml], { type: "text/html" });
+        iframe.src = URL.createObjectURL(blob);
 
         window.addEventListener("message", function handler(event) {
           if (event.data?.type === "parent-access-result") {
@@ -148,7 +104,7 @@ test.describe("GraphSandbox Security", () => {
 
         document.body.appendChild(iframe);
       });
-    });
+    }, generateIframeHtml());
 
     // Verify parent access was blocked
     expect((result as any).blocked).toBe(true);
